generator/requirements.txt
anthropic==0.43.0
pyyaml==6.0.2
